+++ crates/state/src/l1/utils.rs
@@ -1,6 +1,4 @@
-use bitcoin::{
-    block::Header, consensus::Encodable, hashes::Hash, Block, Transaction, WitnessCommitment, Wtxid,
-};
+use bitcoin::{block::Header, consensus::Encodable, hashes::Hash, Block, Wtxid};
 use strata_primitives::{buf::Buf32, hash::sha256d, utils::get_cohashes};
 
 use crate::{
@@ -22,37 +20,6 @@ pub fn compute_block_hash(header: &Header) -> Buf32 {
         .consensus_encode(&mut writer)
         .expect("engines don't error");
     sha256d(&buf)
-}
-
-/// Scans the given coinbase transaction for a witness commitment and returns it if found.
-///
-/// This function iterates over the outputs of the provided `coinbase` transaction from the end
-/// towards the beginning, looking for an output whose `script_pubkey` starts with the “magic” bytes
-/// `[0x6a, 0x24, 0xaa, 0x21, 0xa9, 0xed]`. This pattern indicates an `OP_RETURN` with an
-/// embedded witness commitment header. If such an output is found, the function extracts the
-/// following 32 bytes as the witness commitment and returns a `WitnessCommitment`.
-///
-/// This is based on: [rust-bitcoin](https://github.com/rust-bitcoin/rust-bitcoin/blob/b97be3d4974d40cf348b280718d1367b8148d1ba/bitcoin/src/blockdata/block.rs#L190-L210)
-pub fn witness_commitment_from_coinbase(coinbase: &Transaction) -> Option<WitnessCommitment> {
-    // Consists of OP_RETURN, OP_PUSHBYTES_36, and four "witness header" bytes.
-    const MAGIC: [u8; 6] = [0x6a, 0x24, 0xaa, 0x21, 0xa9, 0xed];
-
-    if !coinbase.is_coinbase() {
-        return None;
-    }
-
-    // Commitment is in the last output that starts with magic bytes.
-    if let Some(pos) = coinbase
-        .output
-        .iter()
-        .rposition(|o| o.script_pubkey.len() >= 38 && o.script_pubkey.as_bytes()[0..6] == MAGIC)
-    {
-        let bytes =
-            <[u8; 32]>::try_from(&coinbase.output[pos].script_pubkey.as_bytes()[6..38]).unwrap();
-        Some(WitnessCommitment::from_byte_array(bytes))
-    } else {
-        None
-    }
 }
 
 /// Generates an L1 transaction with proof for a given transaction index in a block.
